Use predetermined defaults for the config instead of relying on default values (nil)
This is a different approach to what #35 describes. The same config gives hints as to what should be in the config, however the default config should be a bit safer. For instance, the default config has no storage paths - so nothing can actually be uploaded without a config. This is intentional.