Ensure streams are available when downloading remote media
There's a race condition where a stream might not be available if the media is remote and being downloaded. After a couple download attempts, it fixes itself. This will acquire a media stream if one isn't provided to ensure that a stream is always available to the consumer.

Fixes https://github.com/turt2live/matrix-media-repo/issues/142