Strip `charset` off binary media before replying with a `Content-Type`
The spec is a bit unclear whether this is even valid for things like images, however the parameter doesn't really make sense for non-text data anyways.

Synapse currently (for better or worse) treats the presence as a charset as a completely different media type and thus refuses to thumbnail it.