Redis signaling for upload notifier
Expands async upload notify support to multiple processes.

The logic is all instances of media repo subscribe to all shards
for upload notify events and any instance published an event to
one of them. This could be done in reverse as well but I found this
somewhat easier to do.